+++ android/engine/src/test/java/org/smartregister/fhircore/engine/ui/login/LoginViewModelTest.kt
@@ -19,7 +19,11 @@ package org.smartregister.fhircore.engine.ui.login
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
+import io.mockk.every
+import io.mockk.just
+import io.mockk.runs
 import io.mockk.spyk
+import io.mockk.verify
 import javax.inject.Inject
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.junit.After
@@ -27,15 +31,22 @@ import org.junit.Assert
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
+import org.robolectric.annotation.Config
 import org.smartregister.fhircore.engine.app.fakes.FakeModel.authCredentials
 import org.smartregister.fhircore.engine.auth.AccountAuthenticator
+import org.smartregister.fhircore.engine.data.remote.model.response.OAuthResponse
+import org.smartregister.fhircore.engine.robolectric.AccountManagerShadow
 import org.smartregister.fhircore.engine.robolectric.RobolectricTest
 import org.smartregister.fhircore.engine.rule.CoroutineTestRule
 import org.smartregister.fhircore.engine.util.DispatcherProvider
+import org.smartregister.fhircore.engine.util.SecureSharedPreference
 import org.smartregister.fhircore.engine.util.SharedPreferencesHelper
+import retrofit2.Call
+import retrofit2.Response
 
 @ExperimentalCoroutinesApi
 @HiltAndroidTest
+@Config(shadows = [AccountManagerShadow::class])
 internal class LoginViewModelTest : RobolectricTest() {
 
   @get:Rule(order = 0) val hiltRule = HiltAndroidRule(this)
@@ -50,6 +61,8 @@ internal class LoginViewModelTest : RobolectricTest() {
 
   @Inject lateinit var sharedPreferencesHelper: SharedPreferencesHelper
 
+  @Inject lateinit var secureSharedPreference: SecureSharedPreference
+
   private lateinit var loginViewModel: LoginViewModel
 
   private lateinit var accountAuthenticatorSpy: AccountAuthenticator
@@ -61,12 +74,10 @@ internal class LoginViewModelTest : RobolectricTest() {
     accountAuthenticatorSpy = spyk(accountAuthenticator)
 
     loginViewModel =
-      spyk(
-        LoginViewModel(
-          accountAuthenticator = accountAuthenticatorSpy,
-          dispatcher = dispatcherProvider,
-          sharedPreferences = sharedPreferencesHelper
-        )
+      LoginViewModel(
+        accountAuthenticator = accountAuthenticatorSpy,
+        dispatcher = dispatcherProvider,
+        sharedPreferences = sharedPreferencesHelper
       )
   }
 
@@ -88,5 +99,104 @@ internal class LoginViewModelTest : RobolectricTest() {
 
     val successfulLocalLogin = loginViewModel.attemptLocalLogin()
     Assert.assertTrue(successfulLocalLogin)
+  }
+
+  @Test
+  fun testAttemptLocalLoginWithWrongCredentials() {
+    // Simulate saving of credentials prior to login
+    accountAuthenticatorSpy.secureSharedPreference.saveCredentials(authCredentials)
+
+    // Provide username and password (The saved password is hashed, actual one is needed)
+    loginViewModel.run {
+      onUsernameUpdated("hello")
+      onPasswordUpdated("51r1K4l1")
+    }
+
+    val successfulLocalLogin = loginViewModel.attemptLocalLogin()
+    Assert.assertFalse(successfulLocalLogin)
+  }
+
+  @Test
+  fun testAttemptLocalLoginWithNewUser() {
+
+    // Provide username and password (The saved password is hashed, actual one is needed)
+    loginViewModel.run {
+      onUsernameUpdated("demo")
+      onPasswordUpdated("51r1K4l1")
+    }
+
+    val callMock = spyk<Call<OAuthResponse>>()
+
+    every { callMock.enqueue(any()) } just runs
+
+    every { accountAuthenticatorSpy.fetchToken(any(), any()) } returns callMock
+
+    loginViewModel.attemptRemoteLogin()
+
+    // Login error is reset
+    Assert.assertNotNull(loginViewModel.loginError.value)
+    Assert.assertTrue(loginViewModel.loginError.value!!.isEmpty())
+
+    // Show progress bar active
+    Assert.assertNotNull(loginViewModel.showProgressBar.value)
+    Assert.assertTrue(loginViewModel.showProgressBar.value!!)
+
+    verify { accountAuthenticatorSpy.fetchToken(any(), any()) }
+  }
+
+  @Test
+  fun testOauthResponseHandlerHandleSuccessfulResponse() {
+
+    // Provide username and password (The saved password is hashed, actual one is needed)
+    loginViewModel.run {
+      onUsernameUpdated("demo")
+      onPasswordUpdated("51r1K4l1")
+    }
+
+    val callMock = spyk<Call<OAuthResponse>>()
+
+    val mockResponse: Response<OAuthResponse> =
+      Response.success(
+        OAuthResponse(
+          accessToken = authCredentials.sessionToken,
+          tokenType = "openid email profile",
+          refreshToken = authCredentials.refreshToken,
+          scope = "openid"
+        )
+      )
+
+    loginViewModel.oauthResponseHandler.handleResponse(call = callMock, response = mockResponse)
+
+    // Show progress bar inactive
+    Assert.assertNotNull(loginViewModel.showProgressBar.value)
+    Assert.assertFalse(loginViewModel.showProgressBar.value!!)
+
+    // New user credentials added
+    val retrieveCredentials = secureSharedPreference.retrieveCredentials()
+    Assert.assertNotNull(retrieveCredentials)
+    Assert.assertEquals(authCredentials.username, retrieveCredentials!!.username)
+    Assert.assertEquals(authCredentials.sessionToken, retrieveCredentials.sessionToken)
+  }
+
+  @Test
+  fun testForgotPasswordLoadsContact() {
+    loginViewModel.forgotPassword()
+    Assert.assertEquals("tel:0123456789", loginViewModel.launchDialPad.value)
+  }
+
+  @Test
+  fun testAttemptRemoteLoginWithCredentialsCallsAccountAuthenticator() {
+
+    // Provide username and password
+    loginViewModel.run {
+      onUsernameUpdated("testUser")
+      onPasswordUpdated("51r1K4l1")
+    }
+
+    loginViewModel.attemptRemoteLogin()
+
+    Assert.assertEquals("", loginViewModel.loginError.value)
+    loginViewModel.showProgressBar.value?.let { Assert.assertTrue(it) }
+    verify { accountAuthenticatorSpy.fetchToken("testUser", "51r1K4l1".toCharArray()) }
   }
 }