+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/patient/details/QuestPatientDetailViewModel.kt
@@ -17,15 +17,25 @@
 package org.smartregister.fhircore.quest.ui.patient.details
 
 import android.content.Context
+import androidx.annotation.StringRes
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.google.android.fhir.search.search
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 import kotlinx.coroutines.launch
 import kotlinx.serialization.Serializable
+import org.hl7.fhir.r4.model.Bundle
+import org.hl7.fhir.r4.model.Condition
+import org.hl7.fhir.r4.model.Library
+import org.hl7.fhir.r4.model.Observation
+import org.hl7.fhir.r4.model.Patient
+import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
+import org.hl7.fhir.r4.model.Resource
 import org.smartregister.fhircore.engine.configuration.view.SearchFilter
+import org.smartregister.fhircore.engine.cql.LibraryEvaluator
 import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireConfig
 import org.smartregister.fhircore.engine.util.AssetUtil
 import org.smartregister.fhircore.quest.data.patient.PatientRepository
@@ -35,14 +45,17 @@ import org.smartregister.fhircore.quest.ui.patient.register.PatientItemMapper
 @HiltViewModel
 class QuestPatientDetailViewModel
 @Inject
-constructor(val patientRepository: PatientRepository, val patientItemMapper: PatientItemMapper) :
-  ViewModel() {
+constructor(
+  val patientRepository: PatientRepository,
+  val patientItemMapper: PatientItemMapper,
+  val libraryEvaluator: LibraryEvaluator
+) : ViewModel() {
 
   val patientItem = MutableLiveData<PatientItem>()
   val questionnaireConfigs = MutableLiveData<List<QuestionnaireConfig>>()
-  val testResults = MutableLiveData<List<QuestionnaireResponse>>()
+  val testResults = MutableLiveData<List<Pair<QuestionnaireResponse, Questionnaire>>>()
   val onBackPressClicked = MutableLiveData(false)
-  val onMenuItemClicked = MutableLiveData(false)
+  val onMenuItemClicked = MutableLiveData(-1)
   val onFormItemClicked = MutableLiveData<QuestionnaireConfig>(null)
   val onFormTestResultClicked = MutableLiveData<QuestionnaireResponse>(null)
 
@@ -69,8 +82,57 @@ constructor(val patientRepository: PatientRepository, val patientItemMapper: Pat
     viewModelScope.launch { testResults.postValue(patientRepository.fetchTestResults(patientId)) }
   }
 
-  fun onMenuItemClickListener(menuClicked: Boolean) {
-    onMenuItemClicked.value = menuClicked
+  suspend fun getAllDataFor(patientId: String): Bundle {
+    val fhirEngine = patientRepository.fhirEngine
+    val patient = fhirEngine.load(Patient::class.java, patientId)
+    val observations =
+      fhirEngine.search<Observation> {
+        filter(Observation.SUBJECT) { this.value = "Patient/$patientId" }
+      }
+    val conditions =
+      fhirEngine.search<Condition> {
+        filter(Condition.SUBJECT) { this.value = "Patient/$patientId" }
+      }
+
+    val everything =
+      mutableListOf<Resource>().apply {
+        add(patient)
+        addAll(observations)
+        addAll(conditions)
+      }
+
+    return libraryEvaluator.createBundle(everything)
+  }
+
+  fun runCqlFor(patientId: String, context: Context): MutableLiveData<String?> {
+    val result = MutableLiveData("")
+    viewModelScope.launch {
+      val dataBundle = getAllDataFor(patientId)
+      val config =
+        AssetUtil.decodeAsset<ProfileConfig>(fileName = PROFILE_CONFIG, context = context)
+
+      val fhirEngine = patientRepository.fhirEngine
+      val data =
+        kotlin
+          .runCatching {
+            libraryEvaluator
+              .runCqlLibrary(
+                library = fhirEngine.load(Library::class.java, config.cqlProfileLibraryFilter.code),
+                helper = fhirEngine.load(Library::class.java, config.cqlHelperLibraryFilter.code),
+                valueSet = Bundle(),
+                data = dataBundle
+              )
+              .joinToString("\n")
+          }
+          .onFailure { result.postValue(it.stackTraceToString()) }
+          .getOrNull()
+      result.postValue(data)
+    }
+    return result
+  }
+
+  fun onMenuItemClickListener(@StringRes id: Int) {
+    onMenuItemClicked.value = id
   }
 
   fun onFormItemClickListener(questionnaireConfig: QuestionnaireConfig) {
@@ -85,9 +147,19 @@ constructor(val patientRepository: PatientRepository, val patientItemMapper: Pat
     onBackPressClicked.value = backPressed
   }
 
+  fun fetchResultItemLabel(testResult: Pair<QuestionnaireResponse, Questionnaire>): String? {
+    return testResult.second.name?.let { name -> name }
+      ?: testResult.second.title?.let { title -> title }
+  }
+
   companion object {
     const val PROFILE_CONFIG = "configurations/form/profile_config.json"
   }
 
-  @Serializable data class ProfileConfig(val profileQuestionnaireFilter: SearchFilter)
+  @Serializable
+  data class ProfileConfig(
+    val profileQuestionnaireFilter: SearchFilter,
+    val cqlHelperLibraryFilter: SearchFilter,
+    val cqlProfileLibraryFilter: SearchFilter
+  )
 }